refactor(integration): move HttpTransportation to shared java module
Extract the JDK-HttpClient-backed Wirespec.Transportation out of the
kotest integration into a new dependency-light src:integration:java
module (published as community.flock.wirespec.integration:java-jvm), so
other integrations can share it. The module api-exposes the wirespec
runtime since HttpTransportation implements Wirespec.Transportation.

Repackage to community.flock.wirespec.integration.java.transport and
repoint the gradle-kotest example (imports + catalog + dependency) and
the kotest README.
