+++ settings.gradle.kts
@@ -42,6 +42,7 @@ include(
     "src:converter:openapi",
     "src:integration:avro",
     "src:integration:jackson",
+    "src:integration:java",
     "src:integration:kotest",
     "src:integration:kotlinx-serialization",
     "src:integration:wirespec",
