+++ src/integration/java/build.gradle.kts
@@ -1,0 +1,39 @@
+plugins {
+    id("module.publication")
+    id("module.spotless")
+    alias(libs.plugins.kotlin.multiplatform)
+}
+
+group = "${libs.versions.group.id.get()}.integration"
+version = System.getenv(libs.versions.from.env.get()) ?: libs.versions.default.get()
+
+repositories {
+    mavenCentral()
+    mavenLocal()
+}
+
+kotlin {
+    compilerOptions {
+        apiVersion.set(org.jetbrains.kotlin.gradle.dsl.KotlinVersion.fromVersion(libs.versions.kotlin.api.get()))
+        languageVersion.set(org.jetbrains.kotlin.gradle.dsl.KotlinVersion.fromVersion(libs.versions.kotlin.language.get()))
+    }
+    jvm {
+        java {
+            toolchain {
+                languageVersion.set(JavaLanguageVersion.of(libs.versions.java.get()))
+            }
+        }
+    }
+    sourceSets {
+        jvmMain {
+            dependencies {
+                // HttpTransportation implements Wirespec.Transportation, part of this
+                // module's public surface, so the runtime is exposed as api.
+                api(project(":src:integration:wirespec"))
+                // transport() runs the java.net.http call off the calling coroutine
+                // via withContext(Dispatchers.IO).
+                implementation(libs.kotlinx.coroutines.core)
+            }
+        }
+    }
+}
