+++ src/integration/java/src/jvmMain/kotlin/community/flock/wirespec/integration/java/transport/HttpTransportation.kt
@@ -1,4 +1,4 @@
-package community.flock.wirespec.integration.kotest.transport
+package community.flock.wirespec.integration.java.transport
 
 import community.flock.wirespec.kotlin.Wirespec
 import kotlinx.coroutines.Dispatchers
@@ -13,10 +13,9 @@ import java.nio.charset.StandardCharsets
 /**
  * A JDK-`HttpClient`-backed [Wirespec.Transportation]: it turns the generated client's
  * [Wirespec.RawRequest] into a real HTTP call against [baseUrl] and maps the reply back to a
- * [Wirespec.RawResponse]. Pass it to a
- * [WirespecEndpointContext][community.flock.wirespec.integration.kotest.WirespecEndpointContext]
- * so that `SomeEndpoint.Request.call()` exercises a running server over the wire rather than
- * in-process. Dependency-free — only `java.net.http` and coroutines.
+ * [Wirespec.RawResponse]. Wire it into any consumer that drives `SomeEndpoint.Request.call()`
+ * so requests exercise a running server over the wire rather than in-process.
+ * Dependency-free — only `java.net.http` and coroutines.
  *
  * @param baseUrl scheme + host + optional port, e.g. `http://localhost:8080` (no trailing slash).
  */
